project_at_v1_1/user_register/urls.py
from django.conf.urls import patterns, url
from user_register import views
urlpatterns = patterns('', 
	url(r'^$', views.home, name="home"),
	url(r'^welcome/$', views.welcome, name="welcome"),
	url(r'^welcome/choose/$', views.wc, name="wchoose"),
	)