project_at_v1_1/requirements.txt
Coffin==0.3.8
git+https://github.com/django-nonrel/django@nonrel-1.6
git+https://github.com/django-nonrel/mongodb-engine@master
djangotoolbox==1.6.2
pymongo==2.7.2
permission-backend-nonrel==0.2
Pillow==2.6.1
Jinja2==2.7.3
MarkupSafe==0.23
South==1.0
Unidecode==0.04.16
akismet==0.2.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
beautifulsoup4==4.1.3
billiard==3.3.0.18
bleach==1.4
boto==2.31.1
celery==3.1.14
chardet==2.2.1
dj-database-url==0.3.0
dj-static==0.0.5
django-appconf==0.6
django-celery==3.0.11
django-compressor==1.2
django-countries==1.0.5
django-followit==0.0.3
django-keyedcache==1.5.1
django-kombu==0.9.4
django-picklefield==0.3.0
django-recaptcha-works==0.3.4
django-robots==1.0
django-s3-storages-utils==0.1.0
django-storages==1.1.8
django-threaded-multihost==1.4-1
django-tinymce==1.5.1b2
django-toolbelt==0.0.1
djrill==1.2.0
filebrowser-safe==0.3.5
future==0.9.0
grappelli-safe==0.3.12
gunicorn==19.0.0
html5lib==0.90
httplib2==0.9
jsmin==2.0.11
kombu==3.0.22
lamson==1.3.4
lockfile==0.9.1
longerusername==0.4
markdown2==2.2.1
mock==1.0.1
nose==1.3.4
numpy==1.9.0
oauth2==1.5.211
oauthlib==0.6.3
pandas==0.14.1
psycopg2==2.5.3
pystache==0.3.1
python-daemon==1.6.1
python-dateutil==2.2
python-modargs==1.7
python-openid==2.2.5
python-social-auth==0.1.26
pytz==2014.4
requests==2.3.0
requests-oauthlib==0.4.1
sanction==0.3.1
six==1.7.3
static==1.0.2
tzlocal==1.0
wsgiref==0.1.2
